Added Level 2 Tutorial, Made Tutorial Abstract class
