+++ Assets/Scripts/Managers/TutorialManagerLevel1.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 033a89e8d37b7574f94822213134b999
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
