+++ Assets/Scripts/Managers/TutorialManagerLevel2.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e169517fd4aaf8846bfd7db78036ee25
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
